Refactor float_maps example to analyze Taylor-series term counts for f32 instead of f16
