+++ examples/float_maps.rs
@@ -1,72 +1,46 @@
-//! Example: approximate `cos` for `f16` with a plain Taylor series, and use
-//! an exhaustive sweep over every finite `f16` value -- there are only
-//! about 61,000 of them -- to discover how many series terms each one needs
-//! to guarantee an error under one `f32` epsilon (`f32::EPSILON`), for the
-//! values that matter: `x` in `scope` (see `main`).
+//! Example: the `f32` counterpart to `float_maps.rs`, which does the same
+//! Taylor-series term-count analysis for `f16`. There are ~4.28 billion
+//! finite `f32` values -- about 65,000 times as many as `f16`'s ~61,000 --
+//! but this version sweeps every single one anyway, the same brute-force
+//! way `float_maps.rs` does, just to see whether that's actually feasible.
 //!
-//! "ULP" stands for "unit in the last place": the gap between one
-//! representable float and the very next one. It's not a fixed number --
-//! it grows as the magnitude of the value grows, since floats pack more
-//! precision near zero than far from it. This crate's `.next()`/`.prev()`
-//! step by exactly 1 ULP. Here we use a fixed target error of `f32::EPSILON`
-//! (1 ULP near magnitude 1) rather than "1 ULP of whatever `cos(x)` happens
-//! to be": near `cos`'s zero-crossings the true value is tiny, so its own
-//! ULP is tiny too, and demanding that precision would be unreachable.
+//! Unlike `float_maps.rs`, this example needs no nightly feature: `f32` is
+//! a stable type, so `total_float_experimental` (stable) is enough.
 //!
-//! The term count per value is found from the textbook alternating-series
-//! remainder bound, `|x|^(2N) / (2N)!`, rather than by empirically comparing
-//! against a reference `cos` -- that bound is a pure function of `|x|`, so
-//! it can't flicker between "N terms" and "N+1 terms" the way a
-//! rounding-sensitive empirical search would right at the boundary. The
-//! result is captured as a `RangeMapBlaze<FiniteF16, TermsNeeded>`, which
-//! coalesces neighboring `f16` values that need the same term count into a
-//! single, clean, non-overlapping range.
-//!
-//! Why only a limited `scope`? A raw Taylor series only converges quickly
-//! near zero. Real math libraries handle an `x` of any size by first
-//! folding it into a bounded range using `cos`'s periodicity and evenness
-//! (`cos(-x) = cos(x)`, so strictly only `[0, pi]` is ever necessary).
-//! Implementing that fold is a separate concern from the point of this
-//! demo, so it's left out: we simply restrict our attention to `f16` values
-//! already inside `scope` and report everything else as a single
-//! `out of scope` bucket. `scope` is a `RangeSetBlaze<FiniteF16>` built once
-//! in `main` and threaded into `terms_needed`, which decides both "is this
-//! in scope" and "how many terms" together -- so widening or narrowing the
-//! domain (as we've done a few times already) only means editing where
-//! `scope` is built.
-//! The point here is the `RangeMapBlaze`, not the numerics.
-//!
-//! Note: this example only builds with `--features
-//! total_float_nightly_experimental` (see `required-features` in
-//! `Cargo.toml`), so the whole file can assume that feature is on -- no
-//! `#[cfg(...)]` needed anywhere below.
-
-#![feature(f16)]
+//! See `float_maps.rs` for the ULP/epsilon/scope background this example
+//! builds on without re-explaining.
 
 use core::f64::consts::PI;
 use core::num::NonZero;
-use range_set_blaze::{FiniteF16, Integer, RangeMapBlaze, RangeSetBlaze};
+use range_set_blaze::{FiniteF32, Integer, RangeMapBlaze, RangeSetBlaze};
 
 /// Highest number of Taylor terms we're willing to try before giving up.
 const MAX_TERMS: u8 = 30;
-/// Target absolute error: `f32::EPSILON` is 1 ULP near magnitude 1 -- see
-/// the module doc for why a fixed target beats "1 ULP of the reference".
+/// Target absolute error: `f32::EPSILON` is 1 ULP near magnitude 1.
 const TARGET_ERROR: f64 = f32::EPSILON as f64;
 
 fn main() {
-    // Sweep every finite f16 and collect (value, terms-needed) pairs into a
-    // RangeMapBlaze -- it automatically merges neighboring f16 values that
+    // Sweep every finite f32 and collect (value, terms-needed) pairs into a
+    // RangeMapBlaze -- it automatically merges neighboring f32 values that
     // need the same term count into a single range.
-    let all_f16 = !RangeSetBlaze::<FiniteF16>::default(); // idiom for universe of all values
-    let scope = RangeSetBlaze::from_iter([FiniteF16::new(-PI as f16)..=FiniteF16::new(PI as f16)]);
+    let all_f32 = !RangeSetBlaze::<FiniteF32>::default(); // idiom for universe of all values
+    let scope = RangeSetBlaze::from_iter([FiniteF32::new(-PI as f32)..=FiniteF32::new(PI as f32)]);
+    let total = all_f32.len();
 
-    let term_map: RangeMapBlaze<FiniteF16, TermsNeeded> = all_f16
+    let term_map: RangeMapBlaze<FiniteF32, TermsNeeded> = all_f32
         .iter()
-        .map(|x| (x..=x, terms_needed(x, &scope)))
+        .enumerate()
+        .map(|(i, x)| {
+            if i % 100_000_000 == 0 {
+                let pct = 100.0 * i as f64 / f64::from(total);
+                eprintln!("... {} values processed ({pct:.1}%)", with_underscores(i));
+            }
+            (x..=x, terms_needed(x, &scope))
+        })
         .collect();
 
     println!(
-        "Taylor-series terms needed for cos(x) to guarantee an error under one f32 epsilon, for f16 in [-pi, pi]:"
+        "Taylor-series terms needed for cos(x) to guarantee an error under one f32 epsilon, for f32 in [-pi, pi]:"
     );
     for (range, terms) in term_map.range_values() {
         let (start, end) = range.into_inner();
@@ -76,20 +50,35 @@ fn main() {
             TermsNeeded::Terms(n) => format!("{n} terms"),
         };
         println!(
-            "  [{:>12}, {:>12}] -> {label}",
-            start.into_inner() as f32,
-            end.into_inner() as f32,
+            "  [{:>14e}, {:>14e}] -> {label}",
+            start.into_inner(),
+            end.into_inner(),
         );
     }
     println!(
-        "\n{} disjoint ranges cover all {} finite f16 values.",
-        term_map.range_values().count(),
-        term_map.len()
+        "\n{} disjoint ranges cover all {} finite f32 values.",
+        with_underscores(term_map.range_values().count()),
+        with_underscores(term_map.len())
     );
     println!(
-        "Mean terms per in-scope f16 value (each value weighted equally): {:.2}",
+        "Mean terms per in-scope f32 value (each value weighted equally): {:.2}",
         mean_terms(&term_map)
     );
+}
+
+/// Formats a non-negative integer with `_` every three digits (e.g.
+/// `4278190079` -> `4_278_190_079`) -- `std` has no built-in thousands
+/// separator for `Display`.
+fn with_underscores(n: impl core::fmt::Display) -> String {
+    let digits = n.to_string();
+    let mut grouped = String::with_capacity(digits.len() + digits.len() / 3);
+    for (i, digit) in digits.chars().rev().enumerate() {
+        if i != 0 && i % 3 == 0 {
+            grouped.push('_');
+        }
+        grouped.push(digit);
+    }
+    grouped.chars().rev().collect()
 }
 
 /// How many Taylor terms `x` needs, or that `x` is `OutOfScope` (not a
@@ -109,7 +98,7 @@ enum TermsNeeded {
 /// terms it needs: smallest `N` (1..=`MAX_TERMS`) for which the
 /// alternating-series remainder bound, `|x|^(2N) / (2N)!`, guarantees an
 /// error under `TARGET_ERROR`.
-fn terms_needed(x: FiniteF16, scope: &RangeSetBlaze<FiniteF16>) -> TermsNeeded {
+fn terms_needed(x: FiniteF32, scope: &RangeSetBlaze<FiniteF32>) -> TermsNeeded {
     if !scope.contains(x) {
         return TermsNeeded::OutOfScope;
     }
@@ -126,20 +115,20 @@ fn terms_needed(x: FiniteF16, scope: &RangeSetBlaze<FiniteF16>) -> TermsNeeded {
     unreachable!("loop always returns once n reaches MAX_TERMS")
 }
 
-/// Plain arithmetic mean of the term count across every in-scope `f16`
+/// Plain arithmetic mean of the term count across every in-scope `f32`
 /// value, each counted once -- not an average over the reals in `scope`'s
 /// bounds, which would weight every magnitude equally regardless of how
-/// many (or few) `f16` values actually live there. Exact, not estimated:
+/// many (or few) `f32` values actually live there. Exact, not estimated:
 /// `term_map` already holds every in-scope value's term count, coalesced
 /// into a handful of ranges, so this just weights each range's term count
 /// by `Integer::safe_len` (its element count) and takes the weighted
 /// average -- no sampling needed.
-fn mean_terms(term_map: &RangeMapBlaze<FiniteF16, TermsNeeded>) -> f64 {
+fn mean_terms(term_map: &RangeMapBlaze<FiniteF32, TermsNeeded>) -> f64 {
     let mut weighted_sum = 0.0;
     let mut in_scope_count = 0.0;
     for (range, terms) in term_map.range_values() {
         if let TermsNeeded::Terms(n) = terms {
-            let len = FiniteF16::safe_len_to_f64_lossy(FiniteF16::safe_len(&range));
+            let len = FiniteF32::safe_len_to_f64_lossy(FiniteF32::safe_len(&range));
             weighted_sum += len * f64::from(n.get());
             in_scope_count += len;
         }
